Few Changes are made and test cases are added.

+++ app/src/main/java/movierating/models/Movie.java
@@ -32,9 +32,9 @@ public class Movie {
         return genre;
     }
 
-    public int getAverageReviewScore() {
+    public double getAverageReviewScore() {
 
-        return this.getTotalReviewScore() / reviewList.size();
+        return (double) this.getTotalReviewScore() / reviewList.size();
 
     }
